requirements.txt
mcp
openai
python-dotenv
pandas
openpyxl
pyyaml
